+++ template-sample-app/template/src/index.js
@@ -4,17 +4,13 @@
  *  - react router
  *  - main component: App
  */
-import 'react-app-polyfill/stable';
-import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
+import { createCartoSlice, createOauthCartoSlice } from '@carto/react-redux';
 import App from './App';
-
 import { initialState, oauthInitialState } from 'store/initialStateSlice';
 import configureAppStore from './store/store';
-
-import { createCartoSlice, createOauthCartoSlice } from '@carto/react-redux';
 
 const store = configureAppStore();