+++ template-skeleton-cn/template/src/routes.js
@@ -5,9 +5,9 @@ import { useAuth0 } from '@auth0/auth0-react';
 import Header from 'components/common/Header';
 import { initialState } from 'store/initialStateSlice';
 
-const Main = lazy(() => import('components/views/Main'));
+const Main = lazy(() => import(/* webpackPrefetch: true */ 'components/views/main/Main'));
 const NotFound = lazy(() => import('components/views/NotFound'));
-const Login = lazy(() => import('components/views/login/Login'));
+const Login = lazy(() => import('components/views/Login'));
 // [hygen] Import views
 
 export const ROUTE_PATHS = {