+++ template-skeleton/template/src/components/common/Header.js
@@ -12,14 +12,18 @@ import {
   Link,
   makeStyles,
   Typography,
+  Avatar,
+  Menu,
+  MenuItem,
 } from '@material-ui/core';
 import MenuIcon from '@material-ui/icons/Menu';
 import CloseIcon from '@material-ui/icons/Close';
-import UserMenu from 'components/views/login/UserMenu';
 import { NavLink, useLocation } from 'react-router-dom';
-import { ROUTE_PATHS } from 'routes';
 import { ReactComponent as CartoLogo } from 'assets/img/carto-logo.svg';
 import { ReactComponent as CartoLogoXS } from 'assets/img/carto-logo-xs.svg';
+import { useDispatch, useSelector } from 'react-redux';
+import { logout } from '@carto/react-redux';
+import { ROUTE_PATHS } from 'routes';
 
 const useStylesCommon = makeStyles((theme) => ({
   title: {
@@ -80,7 +84,7 @@ function Desktop() {
 
   return (
     <Hidden xsDown>
-      <Link component={NavLink} to='/' className={classes.title}>
+      <Link component={NavLink} to={ROUTE_PATHS.DEFAULT} className={classes.title}>
         <Typography component='h1' variant='subtitle1' noWrap>
           <CartoLogo />
           <AppName />
@@ -136,7 +140,7 @@ function Mobile() {
         {drawerOpen ? <CloseIcon /> : <MenuIcon />}
       </IconButton>
       <Divider orientation='vertical' light />
-      <Link component={NavLink} to='/' className={classes.title}>
+      <Link component={NavLink} to={ROUTE_PATHS.DEFAULT} className={classes.title}>
         <Typography component='h1' variant='subtitle1' noWrap>
           <Hidden smUp>
             <CartoLogoXS />
@@ -205,5 +209,96 @@ function NavigationMenu({ column: vertical }) {
         {/* [hygen] Import links */}
       </Tabs>
     </Grid>
+  );
+}
+
+const useStylesUserMenu = makeStyles((theme) => ({
+  avatar: {
+    cursor: 'pointer',
+    width: theme.spacing(4.5),
+    height: theme.spacing(4.5),
+    marginLeft: theme.spacing(1),
+  },
+}));
+
+function UserMenu() {
+  const dispatch = useDispatch();
+  const oauthApp = useSelector((state) => state.oauth.oauthApp);
+  const user = useSelector((state) => state.oauth.userInfo);
+  const [anchorEl, setAnchorEl] = useState(null);
+  const classes = useStylesUserMenu();
+
+  // If no OAuthApp has been configured, no user-related controls are displayed
+  // or
+  // User is NOT logged in, so display nothing
+  if (!oauthApp || !user) {
+    return null;
+  }
+
+  // At this point, there is an oauthApp and the user has logged in (forceOAuthLogin mode).
+  const open = Boolean(anchorEl);
+
+  const handleMenu = (event) => {
+    if (!anchorEl) {
+      setAnchorEl(event.currentTarget);
+    } else {
+      setAnchorEl(null);
+    }
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  const handleLogout = () => {
+    dispatch(logout());
+    handleClose();
+  };
+
+  const goToCarto = () => {
+    const url = user.api_endpoints.builder;
+    window.open(url);
+  };
+
+  // Display User menu, with name, avatar + an attached menu for user-related options
+  return (
+    <>
+      <Link
+        edge='end'
+        aria-label='account of current user'
+        aria-controls='menu-login'
+        aria-haspopup='true'
+        onClick={handleMenu}
+        color='inherit'
+      >
+        <Grid container alignItems='center' item wrap='nowrap'>
+          <Hidden smDown>
+            <Typography variant='caption' color='inherit' noWrap>
+              {user.username}
+            </Typography>
+          </Hidden>
+          <Avatar className={classes.avatar} src={user.avatar_url} />
+        </Grid>
+      </Link>
+      <Menu
+        id='menu-login'
+        anchorEl={anchorEl}
+        getContentAnchorEl={null}
+        anchorOrigin={{
+          vertical: 'bottom',
+          horizontal: 'right',
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: 'top',
+          horizontal: 'right',
+        }}
+        open={open}
+        onClose={handleClose}
+      >
+        <MenuItem onClick={handleLogout}>Logout</MenuItem>
+        <MenuItem onClick={goToCarto}>Go to CARTO</MenuItem>
+      </Menu>
+    </>
   );
 }